+++ js/index.js
@@ -1,0 +1,7 @@
+import publications from './publications';
+import releases from './releases';
+import blog from './blog';
+
+publications();
+releases();
+blog();
